large_app/python/requirements.txt
apispec==3.0.0
Flask==1.1.1
flask-apispec==0.8.3
Flask-Login==0.4.1
flask-marshmallow==0.10.1
Flask-Migrate==2.4.0
Flask-SQLAlchemy==2.4.0
flask-swagger-ui==3.20.9
loggers==0.1.4
marshmallow==3.2.1
marshmallow-sqlalchemy==0.19.0
python-json-logger==0.1.11
sentry-sdk==0.12.3
SQLAlchemy==1.3.8
python-jose==3.0.1
requests==2.22.0
pymysql==0.9.3
psycopg2==2.8.3